# cuda4dnn(conv): autotuning for convolution