**Statistics as of May 14th 2020:**

Devices:

- GTX 1050 Mobile
- GTX 1080 Ti
- RTX 2080 Ti

[benchmarks without this PR](https://gist.github.com/YashasSamaga/4640594be2580d85b80f5c662339ff4b)

[benchmarks with this PR](https://gist.github.com/YashasSamaga/be44243671b4796977b475d02f41ef80)

[autotuning algorithm selections](https://gist.github.com/YashasSamaga/963d628a10711c8b5bfd6d8da2a5fee0)

[convolution configurations](https://gist.github.com/YashasSamaga/b048ee7084e52d9c9a3dd41f15663e09)

---

**NOTE:** Not up-to-date.

**Device:** GTX 1050

Model                                       | without this patch | with this patch
---------------------------------------- | ----------------------- | ---------------------
MobileNet SSD Coco v1          | 55ms                    |  **7ms**
MobileNet SSD Coco v2          | 84ms                    | **11ms**
OpenPose pose MPI                | 156ms                 | **111ms**
EfficientNet B0 YOLOv3          | 121ms                  | **15ms**
FastNeuralStyle Stary Night    | 26ms                    | **22ms**
Inception v2 Mask RCNN        | 180ms                  | **152ms**

Every model which uses depthwise convolutions has got an improvement.

Most of Mask RCNN improvement comes from a single convolution layer where the algorithm chosen by heuristics took 40ms whereas the best algorithm took just 20ms. 

---

**Pending:**
- insanely big initialization times

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under OpenCV (BSD) License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [x] There is reference to original bug report and related work
- [x] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [x] The feature is well documented and sample code can be built with the project CMake

```
force_builders=Custom
buildworker:Custom=linux-4
build_image:Custom=ubuntu-cuda:18.04
```