@YashasSamaga 2020-03-26
The CUDA backend now performs autotuning on its own which appears have resolved the performance degradation problems. Moreover, it also includes fused convolutions while autotuning. This has improved the performance further.

### Problem:
The initialization time has gone up significantly. It's in several seconds for many models and few of them take two digit seconds. Mask RCNN takes 26s for initialization!

There are eight algorithms available for convolution. These are the combinations which are tried while autotuning:

1. **cuDNN default math fused:**
    - all eight algorithms are tried
    - cuDNN does convolution, bias addition and activation together
2. **cuDNN default math unfused:**
    - all eight algorithms are tried
    - cuDNN does the convolution
    - bias addition and activation are carried out by cuda4dnn kernels (bias and activation are fused together)
3. **cuDNN tensor core fused:**
    - two algorithms are tried
    - cuDNN does convolution, bias addition and activation together
4. **cuDNN tensor core unfused**
    - two algorithms are tried
    - cuDNN does the convolution
    - bias addition and activation are carried out by cuda4dnn kernels (bias and activation are fused together)

In the worst case, 20 combinations will be tried for every convolution layer. There are few algorithms, particularly FFT based algorithms, which request insane amounts of workspace. For example, the FFT tiling algorithm requests a workspace of 8GB for convolving on a 200KB image. These memory allocations take seconds!

The FFT algorithms are feebly used (only once in my testbed consisting of 26 models) as they are too slow for small filters and small batches.

**Why try unfused cuDNN combinations when fused cuDNN combinations are available?**
Because fused cuDNN can sometimes be slower than unfused cuDNN convolution with separate bias_activation step. For example, if fused cuDNN operations are forced whenever they are available, MobileNet to take 37ms instead of just 7ms!