@alalek 2020-03-26
Some similar approach exists in OpenCL (ocl4dnn) backend. It includes:
- rules to define convolution [configuration](https://github.com/opencv/opencv/blob/4.2.0/modules/dnn/src/ocl4dnn/src/ocl4dnn_conv_spatial.cpp#L631-L643).
- [default "pre-tuned" values](https://github.com/opencv/opencv/blob/4.2.0/modules/dnn/src/ocl4dnn/include/default_kernel_config.hpp) for some cases (target H/W is selected through number of execution units (EU) - autotuning is available for Intel iGPUs).
- multi-level storage of tuned configurations: ["default"](https://github.com/opencv/opencv/blob/4.2.0/modules/dnn/src/ocl4dnn/src/ocl4dnn_conv_spatial.cpp#L1973-L1974), ["in-memory"](https://github.com/opencv/opencv/blob/4.2.0/modules/dnn/src/ocl4dnn/src/ocl4dnn_conv_spatial.cpp#L1976), ["on disk"](https://github.com/opencv/opencv/blob/4.2.0/modules/dnn/src/ocl4dnn/src/ocl4dnn_conv_spatial.cpp#L2045).
- tuning configuration [flags](https://github.com/opencv/opencv/blob/4.2.0/modules/dnn/src/ocl4dnn/src/ocl4dnn_conv_spatial.cpp#L204-L233). Auto-tuning is disabled by default, but can be enabled by request or loaded from the disk.
- some [magic](https://github.com/opencv/opencv/blob/4.2.0/modules/dnn/src/ocl4dnn/src/ocl4dnn_conv_spatial.cpp#L1665-L1695) how candidates are generated for tuning process.