@YashasSamaga 2020-05-11
What about having an API for loading, saving and optimizing model cofigurations that is exposed through `cv::dnn::Net`? This seems more user-friendly than using environment variables. It also allows having unoptimized and optimized models to be used simultaneously (without having to modify the environment variables at runtime). It's also easier to extend in the future.

Both OpenCL and CUDA could use the same API instead of having individual backend-specific APIs.

I think only three functions are required (and maybe user-friendly overloads):
```
void Net::loadOptimizedConfiguration(const std::string& config);
void Net::saveOptimizedConfiguration(std::string& config);
void Net::runOptimizer();
```

The configuration could be stored in protobuf format. The format can be kept as an internal format.