@YashasSamaga 2020-07-28
cuDNN 8 has a brand new API which offers better autotuning capabilities. I will make a PR supporting cuDNN 8's new backend API and then a PR for autotuning.